+++ crates/swc/tests/tsc-references/destructuringSpread_es2015.1.normal.js
@@ -1,5 +1,6 @@
 import _object_spread from "@swc/helpers/lib/_object_spread.js";
-const { x  } = _object_spread({}, {}, {
+import _object_spread_props from "@swc/helpers/lib/_object_spread_props.js";
+const { x  } = _object_spread_props(_object_spread({}, {}), {
     x: 0
 });
 const { y  } = _object_spread({
@@ -11,12 +12,12 @@ const { z , a , b  } = _object_spread({
     a: 0,
     b: 0
 });
-const { c , d , e , f , g  } = _object_spread({}, _object_spread({}, _object_spread({}, {
+const { c , d , e , f , g  } = _object_spread_props(_object_spread({}, _object_spread_props(_object_spread({}, _object_spread_props(_object_spread({}, {
     c: 0
-}, {
+}), {
     d: 0
-}), {
+})), {
     e: 0
-}), {
+})), {
     f: 0
 });