+++ crates/swc/tests/tsc-references/jsxJsxsCjsTransformKeyPropCustomImportPragma_es5.1.normal.js
@@ -1,4 +1,5 @@
 import _object_spread from "@swc/helpers/lib/_object_spread.js";
+import _object_spread_props from "@swc/helpers/lib/_object_spread_props.js";
 import { jsx as _jsx } from "preact/jsx-runtime";
 import { createElement as _createElement } from "react";
 // @filename: react.tsx
@@ -12,18 +13,18 @@ import { createElement as _createElement } from "react";
 /* @jsxImportSource preact */ var props = {
     answer: 42
 };
-var a = /*#__PURE__*/ _jsx("div", _object_spread({}, props, {
+var a = /*#__PURE__*/ _jsx("div", _object_spread_props(_object_spread({}, props), {
     children: "text"
 }), "foo");
-var b = /*#__PURE__*/ _createElement("div", _object_spread({}, props, {
+var b = /*#__PURE__*/ _createElement("div", _object_spread_props(_object_spread({}, props), {
     key: "bar"
 }), "text");
 var props2 = {
     answer: 42
 };
-var a2 = /*#__PURE__*/ _jsx("div", _object_spread({}, props2, {
+var a2 = /*#__PURE__*/ _jsx("div", _object_spread_props(_object_spread({}, props2), {
     children: "text"
 }), "foo");
-var b2 = /*#__PURE__*/ _createElement("div", _object_spread({}, props2, {
+var b2 = /*#__PURE__*/ _createElement("div", _object_spread_props(_object_spread({}, props2), {
     key: "bar"
 }), "text");