+++ crates/swc/tests/tsc-references/objectSpreadComputedProperty_es2015.1.normal.js
@@ -1,20 +1,21 @@
 import _object_spread from "@swc/helpers/lib/_object_spread.js";
+import _object_spread_props from "@swc/helpers/lib/_object_spread_props.js";
 // fixes #12200
 function f() {
     let n = 12;
     let m = 13;
     let a = null;
-    const o1 = _object_spread({}, {}, {
+    const o1 = _object_spread_props(_object_spread({}, {}), {
         [n]: n
     });
-    const o2 = _object_spread({}, {}, {
+    const o2 = _object_spread_props(_object_spread({}, {}), {
         [a]: n
     });
-    const o3 = _object_spread({
+    const o3 = _object_spread_props(_object_spread(_object_spread_props(_object_spread({
         [a]: n
-    }, {}, {
+    }, {}), {
         [n]: n
-    }, {}, {
+    }), {}), {
         [m]: m
     });
 }