+++ crates/swc_ecma_preset_env/tests/fixtures/transform/tp-version/input.mjs
@@ -1,0 +1,5 @@
+class A {
+    static {
+        this.abc = 123;
+    }
+}