+++ crates/swc_ecma_transforms_compat/src/es2015/typeof_symbol.rs
@@ -1,8 +1,7 @@
 use swc_atoms::js_word;
 use swc_common::{util::take::Take, DUMMY_SP};
 use swc_ecma_ast::*;
-use swc_ecma_transforms_base::{helper, perf::Parallel};
-use swc_ecma_transforms_macros::parallel;
+use swc_ecma_transforms_base::helper;
 use swc_ecma_utils::{quote_str, ExprFactory};
 use swc_ecma_visit::{as_folder, noop_visit_mut_type, Fold, VisitMut, VisitMutWith};
 use swc_trace_macro::swc_trace;
@@ -16,16 +15,6 @@ pub fn typeof_symbol() -> impl VisitMut + Fold {
 struct TypeOfSymbol;
 
 #[swc_trace]
-impl Parallel for TypeOfSymbol {
-    fn merge(&mut self, _: Self) {}
-
-    fn create(&self) -> Self {
-        TypeOfSymbol
-    }
-}
-
-#[swc_trace]
-#[parallel]
 impl VisitMut for TypeOfSymbol {
     noop_visit_mut_type!();